Added validation for fields and confirmations for DA operations. As well as added a method for random password generation

+++ PasswordCatacomb/Form1.cs
@@ -4,6 +4,7 @@ using System.ComponentModel;
 using System.Data;
 using System.Drawing;
 using System.Linq;
+using System.Security.Cryptography;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
@@ -14,9 +15,8 @@ namespace PasswordCatacomb
 {
     public partial class Form1 : Form
     {
-        public const string constuser = "leifj";
-
         private PasswordInfo _passwordInfo = new PasswordInfo();
+        private List<string> _passwordNamesForUser = new List<string>();
 
         public PasswordInfo PasswordInfo
         {
@@ -24,10 +24,18 @@ namespace PasswordCatacomb
             set { _passwordInfo = value; }
         }
 
-        public Form1()
+        public List<string> PasswordNamesForUser
+        {
+            get { return _passwordNamesForUser; }
+            set { _passwordNamesForUser = value; }
+        }
+
+        public Form1(string user)
         {
             InitializeComponent();
-            PasswordInfo.User = constuser;
+            PasswordInfo.User = user;
+            PasswordNamesForUser = PasswordDA.GetPasswordNamesForUser(user);
+            passwordNameComboBox.DataSource = PasswordNamesForUser;
         }
 
         //private void passwordNameTextBox_TextChanged(object sender, EventArgs e)
@@ -42,6 +50,11 @@ namespace PasswordCatacomb
 
         private void radioButton_CheckedChanged(object sender, EventArgs e)
         {
+            passwordNameTextBox.Text = string.Empty;
+            passwordStringTextBox.Text = string.Empty;
+            passwordNameComboBox.Text = string.Empty;
+            autoGenerateCheckBox.Checked = false;
+
             RadioButton radioButton = sender as RadioButton;
 
             if (radioButton != null && radioButton.Checked)
@@ -49,49 +62,111 @@ namespace PasswordCatacomb
                 if (radioButton == addRadioButton)
                 {
                     enterButton.Text= "Add";
+                    passwordStringTextBox.ReadOnly = false;
+                    passwordNameComboBox.Visible = false;
+                    autoGenerateCheckBox.Visible = true;
+                    autoGenerateLabel.Visible = true;
                 }
                 else if (radioButton == updateRadioButton)
                 {
                     enterButton.Text = "Update";
+                    passwordStringTextBox.ReadOnly = false;
+                    passwordNameComboBox.Visible = true;
+                    passwordNameTextBox.Text = string.Empty;
+                    autoGenerateCheckBox.Visible = true;
+                    autoGenerateLabel.Visible = true;
                 }
                 else if (radioButton == retrieveRadioButton)
                 {
                     enterButton.Text = "Retrieve";
+                    passwordStringTextBox.ReadOnly= true;
+                    passwordNameComboBox.Visible = true;
+                    passwordNameTextBox.Text = string.Empty;
+                    autoGenerateCheckBox.Visible = false;
+                    autoGenerateLabel.Visible = false;
                 }
                 else if (radioButton == deleteRadioButton)
                 {
                     enterButton.Text = "Delete";
+                    passwordStringTextBox.ReadOnly = true;
+                    passwordNameComboBox.Visible = true;
+                    passwordNameTextBox.Text = string.Empty;
+                    autoGenerateCheckBox.Visible = false;
+                    autoGenerateLabel.Visible = false;
                 }
             }
         }
 
         private void enterButton_Click(object sender, EventArgs e)
         {
-            if (enterButton.Text== "Add") 
+            if (enterButton.Text == "Add")
             {
+                if ((string.IsNullOrEmpty(passwordNameTextBox.Text) && string.IsNullOrEmpty(passwordNameComboBox.Text)) || string.IsNullOrEmpty(passwordStringTextBox.Text))
+                {
+                    MessageBox.Show("Please enter both the password name and password.", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    return;
+                }
+
                 PasswordInfo.AESKey = AesHelper.GenerateKey(PasswordInfo.PasswordString);
                 PasswordInfo.AESIV = AesHelper.GenerateIV();
                 PasswordInfo.EncryptedPassword = AesHelper.EncryptStringToBytes(PasswordInfo.PasswordString, PasswordInfo.AESKey, PasswordInfo.AESIV);
                 PasswordDA.CreateRecord(PasswordInfo);
+                passwordNameComboBox.DataSource = PasswordDA.GetPasswordNamesForUser(PasswordInfo.User);
+                passwordNameComboBox.Refresh();
             }
-            if (enterButton.Text == "Update")
+            else if (enterButton.Text == "Update")
             {
+                if ((string.IsNullOrEmpty(passwordNameTextBox.Text) && string.IsNullOrEmpty(passwordNameComboBox.Text)) || string.IsNullOrEmpty(passwordStringTextBox.Text))
+                {
+                    MessageBox.Show("Please enter both the password name and password.", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    return;
+                }
+
                 PasswordInfo.AESKey = AesHelper.GenerateKey(PasswordInfo.PasswordString);
                 PasswordInfo.AESIV = AesHelper.GenerateIV();
                 PasswordInfo.EncryptedPassword = AesHelper.EncryptStringToBytes(PasswordInfo.PasswordString, PasswordInfo.AESKey, PasswordInfo.AESIV);
                 PasswordDA.UpdateRecord(PasswordInfo);
             }
-            if (enterButton.Text == "Retrieve")
+            else if (enterButton.Text == "Retrieve")
             {
+                if (string.IsNullOrEmpty(passwordNameTextBox.Text) && string.IsNullOrEmpty(passwordNameComboBox.Text))
+                {
+                    MessageBox.Show("Please enter the password name.", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    return;
+                }
+
                 PasswordInfo = PasswordDA.ReadRecord(PasswordInfo.User, PasswordInfo.PasswordName);
-                PasswordInfo.PasswordString = AesHelper.DecryptStringFromBytes(PasswordInfo.EncryptedPassword, PasswordInfo.AESKey, PasswordInfo.AESIV);
-                passwordStringTextBox.Text = PasswordInfo.PasswordString;
+                if (PasswordInfo.User != null)
+                {
+                    PasswordInfo.PasswordString = AesHelper.DecryptStringFromBytes(PasswordInfo.EncryptedPassword, PasswordInfo.AESKey, PasswordInfo.AESIV);
+                    passwordStringTextBox.Text = PasswordInfo.PasswordString;
+                }
+                else
+                {
+                    MessageBox.Show("Password with that name does not exist, please try again.", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                }
             }
-            if (enterButton.Text == "Delete")
+            else if (enterButton.Text == "Delete")
             {
+                if (string.IsNullOrEmpty(passwordNameTextBox.Text) && string.IsNullOrEmpty(passwordNameComboBox.Text))
+                {
+                    MessageBox.Show("Please enter the password name.", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    return;
+                }
+
                 PasswordDA.DeleteRecord(PasswordInfo.User, PasswordInfo.PasswordName);
+                passwordNameComboBox.DataSource = PasswordDA.GetPasswordNamesForUser(PasswordInfo.User);
+                passwordNameComboBox.Refresh();
+            }
+
+            if (enterButton.Text != "Retrieve")
+            {
+                passwordNameTextBox.Text = string.Empty;
+                passwordStringTextBox.Text = string.Empty;
+                passwordNameComboBox.Text = string.Empty;
             }
         }
+
 
         private void passwordNameTextBox_Leave(object sender, EventArgs e)
         {
@@ -101,6 +176,44 @@ namespace PasswordCatacomb
         private void passwordStringTextBox_Leave(object sender, EventArgs e)
         {
             PasswordInfo.PasswordString = passwordStringTextBox.Text;
+        }
+
+        private void passwordNameComboBox_Leave(object sender, EventArgs e)
+        {
+            PasswordInfo.PasswordName = passwordNameComboBox.Text;
+        }
+
+        public static string GenerateStrongPassword()
+        {
+            const string chars = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789!@#$%^&*()_+=-{}[]\\|:;<>,.?/";
+            int length = 15;
+
+            using (RNGCryptoServiceProvider crypto = new RNGCryptoServiceProvider())
+            {
+                byte[] data = new byte[length];
+                crypto.GetBytes(data);
+
+                StringBuilder result = new StringBuilder(length);
+                for (int i = 0; i < length; i++)
+                {
+                    int index = data[i] % chars.Length;
+                    result.Append(chars[index]);
+                }
+
+                return result.ToString();
+            }
+        }
+
+        private void autoGenerateCheckBox_CheckedChanged(object sender, EventArgs e)
+        {
+            if (autoGenerateCheckBox.Checked)
+            {
+                PasswordInfo.PasswordString = passwordStringTextBox.Text = GenerateStrongPassword();
+            }
+            else
+            {
+                PasswordInfo.PasswordString = passwordStringTextBox.Text = string.Empty;
+            }
         }
     }
 }
